Add support for the AD849x series of thermocouple amplifiers
New support for heater or temperature blocks configured with
sensor_type of AD8494, AD8495, AD8496, and AD8497.

The reading method uses the simple voltage-temperature linearity
correction table described in AN-1087 for that series of products.

For voltages not covered by the linearity correction tables the code
falls back to using the 5mV per degree linear transfer function
described in the datasheet.

https://www.analog.com/media/en/technical-documentation/application-notes/AN-1087.PDF

Signed-off-by: Geoff Shannon <geoffpshannon@gmail.com>